some splitting of workspaces file into smaller one func per file refactoring to allow code reuse between some files
