+++ packages/hasher/package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@microsoft/rush-lib": "^5.23.2",
     "@pnpm/lockfile-file": "^3.0.7",
+    "@pnpm/logger": "^3.2.2",
     "@yarnpkg/lockfile": "^1.1.0",
     "backfill-config": "^4.0.0",
     "backfill-logger": "^4.0.0",
