+++ packages/hasher/src/__tests__/resolveDependenciesHelper.ts
@@ -1,8 +1,8 @@
 import { setupFixture } from "backfill-utils-test";
 
-import { getYarnWorkspaces } from "../yarn/yarnWorkspaces";
-import { getPnpmWorkspaces } from "../pnpm/pnpmWorkspaces";
-import { getRushWorkspaces } from "../rush/rushWorkspaces";
+import { getYarnWorkspaces } from "../workspaces/yarnWorkspaces";
+import { getPnpmWorkspaces } from "../workspaces/pnpmWorkspaces";
+import { getRushWorkspaces } from "../workspaces/rushWorkspaces";
 
 export async function filterDependenciesInYarnFixture(
   fixture: string,
