+++ packages/hasher/src/__tests__/resolveExternalDependencies.test.ts
@@ -1,14 +1,14 @@
 import { setupFixture } from "backfill-utils-test";
 
-import { getYarnWorkspaces } from "../yarn/yarnWorkspaces";
+import { getYarnWorkspaces } from "../workspaces/yarnWorkspaces";
 import {
   filterExternalDependencies,
   resolveExternalDependencies,
   addToQueue
 } from "../resolveExternalDependencies";
 import { parseLockFile } from "../lockfile";
-import { getPnpmWorkspaces } from "../pnpm/pnpmWorkspaces";
-import { getRushWorkspaces } from "../rush/rushWorkspaces";
+import { getPnpmWorkspaces } from "../workspaces/pnpmWorkspaces";
+import { getRushWorkspaces } from "../workspaces/rushWorkspaces";
 import { filterDependenciesInYarnFixture } from "./resolveDependenciesHelper";
 
 describe("filterExternalDependencies()", () => {
