+++ packages/hasher/src/workspaces/WorkspaceInfo.ts
@@ -1,0 +1,1 @@
+export type WorkspaceInfo = { name: string; path: string }[];
