+++ packages/hasher/src/workspaces/findWorkspacePath.ts
@@ -1,0 +1,12 @@
+import { WorkspaceInfo } from "./WorkspaceInfo";
+
+export function findWorkspacePath(
+  workspaces: WorkspaceInfo,
+  packageName: string
+): string | undefined {
+  const workspace = workspaces.find(({ name }) => name === packageName);
+
+  if (workspace) {
+    return workspace.path;
+  }
+}
