+++ packages/hasher/src/workspaces/listOfWorkspacePackageNames.ts
@@ -1,0 +1,7 @@
+import { WorkspaceInfo } from "./WorkspaceInfo";
+
+export function listOfWorkspacePackageNames(
+  workspaces: WorkspaceInfo
+): string[] {
+  return workspaces.map(({ name }) => name);
+}
